+++ README.md
@@ -96,20 +96,6 @@ docker compose -p kitchenpos up -d
 
 ## 용어 사전
 
-### 메뉴 (Bounded Context)
-
-| 한글명 | 영문명 | 설명 |
-| --- | --- | --- |
-| 메뉴 | menu | 한 개 이상의 메뉴 상품으로 구성된다.<br/>ex) 후라이드 두마리 세트 : 후라이드 치킨 2마리(메뉴상품) + 콜라 1.25L(메뉴상품) |
-| 메뉴 상품 | menuProduct | 어떤 종류의 상품을 몇개를 포함하는지를 의미한다. 메뉴 상품은 하나 이상의 상품을 포함한다.<br/>ex) 후라이드 치킨 2마리 : 후라이드 치킨(상품) x 2(수량) |
-| 메뉴명 | menuName | 메뉴의 이름을 의미한다. |
-| 비속어 | profanity | 욕설 또는 비속어로 손님들이 불편함을 느껴 메뉴명에 포함될 수 없는 단어를 의미한다. |
-| 메뉴 가격 | menuPrice | 메뉴의 가격을 의미한다. |
-| 전시한다 | display | 메뉴를 주문할 수 있도록 정보를 보여준다. |
-| 숨긴다 | hide | 메뉴를 주문할 수 없도록 정보를 숨긴다. |
-| 메뉴그룹 | menuGroup | 메뉴들을 분류하기 위한 그룹으로, 여러 메뉴들을 포함하고 있다.<br/>ex) 추천메뉴 그룹:<br/> - 후라이드 두마리 메뉴<br/> - 마늘, 간장 메뉴 등 |
-| 메뉴그룹명 | menuGroupName | 메뉴 그룹의 이름을 의미한다. |
-
 ### 매장 주문 (Bounded Context)
 
 | 한글명 | 영문명 | 설명 |
@@ -181,29 +167,6 @@ docker compose -p kitchenpos up -d
 | 완료한다 | complete | 주문을 완료한다. |
 
 ## 모델링
-
-### 메뉴
-
-- `메뉴(menu)`는 식별자, 이름, 가격, 전시 여부, 메뉴 상품목록을 가진다.<br/><br/>
-- `메뉴(menu)`를 생성한다.
-  - `메뉴(menu)`는 `메뉴 그룹(menuGroup)`에 포함되어야 한다.
-  - `메뉴(menu)`는 `메뉴 상품(menuProduct)`을 하나 이상 포함해야 한다.
-  - `메뉴명(menuName)`은 `비속어(profanity)`가 포함될 수 없다.
-  - `메뉴 가격(menuPrice)`은 0원 이상이어야 한다.
-  - `메뉴 가격(menuPrice)`은 `메뉴 상품(menuProduct)`의 가격 총합보다 작아야한다.
-- `메뉴(menu)`의 가격을 변경한다.
-  - `메뉴 가격(menuPrice)`은 0원 이상이어야 한다.
-  - `메뉴 가격(menuPrice)`은 `메뉴 상품(menuProduct)`의 가격 총합보다 작아야한다.
-  - `메뉴 가격(menuPrice)`을 변경할 때, 변경 후의 `메뉴 가격(menuPrice)`이 `메뉴 상품(menuProduct)`의 가격의 합보다 크면 `숨겨진다(hide).`
-- `메뉴(menu)`를 `전시한다(display).`
-  - `메뉴 가격(menuPrice)`이 `메뉴 상품(menuProduct)`의 가격 총합보다 높을 경우 `메뉴(menu)`를 `전시할 수 없다(display).`
-- `메뉴(menu)`를 `숨긴다(hide).`
-
-### 메뉴 그룹
-
-- `메뉴그룹(menuGroup)`은 식별자, 이름을 가진다.<br/><br/>
-- `메뉴그룹(menuGroup)`을 생성한다.
-  - `메뉴그룹명(menuGroupName)`이 공백이어서는 안된다.
 
 ### 매장 주문
