+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,6 +1,8 @@
 package kitchenpos.product.application;
 
-import kitchenpos.domain.*;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.infra.PurgomalumClient;
